Tests no longer need to patch stack_ide_loadtargets
The patching was to avoid actually running stack-ide, but
it made tests harder to read and wasn't done consistently.

Since StackIDE class already has some logic to detect when
it is being tested, we do the patching there. Moreover,
we allow customization of the return value via the
FakeBackend class.