Assign StackIDE instances to views, not windows
This patch changes fundamentally the way the StackIDEManager works.
Before this, each sublime window would have a unique associated StackIDE
instance for all its views (tabs, etc). Moreoever, the instance would
be run from the directory associated with the window, which means that
the plugin would only work if the the window was launch from the command
line using `subl .` or something along those lines. This also meant that
we couldn't work with tabs from different projects on the same window.

In this version, we change two things. First, instances are associated
with cabal projects (see the StackIDEInstanceKey class) so we have
one instance per project, globally accross all windows. Second, for
each view we determine its parent cabal project and use that as a
key to get its corresponding StackIDE instance.

Because of this, it no longer matters how we run sublime, the plugin
will just work. More importantly, we can open files from different
projects on the same window and each tab will be associated with
the right appropriate ide-backend instance. This is particularly
useful when one is working on a multi-project using stack, that is,
when one has more than one project mentioned in the packages
entry of stack.yaml).

Since with this design we now know, for each cabal project, which
files are being seen by the user, every time a file is save, we
will only reload those files that occur on some view. This means,
in particular, that errors on files that are not open on some window
will not trigger a new tab to open for that file. This can be a
controversial change but, in my experience, this is a better
behaviour: when one is going through a complex refactoring on a
large project, one knows there will be errors and having sublime
open tabs that we didn't ask for and changing the focus can be
more annyoing than helpful.